F_T30603 move AssessmentTable logic under core
- adjust use statements
