+++ demosplan/DemosPlanCoreBundle/Logic/AssessmentTable/AssessmentTableServiceStorage.php
@@ -8,7 +8,7 @@
  * All rights reserved
  */
 
-namespace demosplan\DemosPlanAssessmentTableBundle\Logic;
+namespace demosplan\DemosPlanCoreBundle\Logic\AssessmentTable;
 
 use BadMethodCallException;
 use Carbon\Carbon;
@@ -45,6 +45,7 @@ use Exception;
 use FOS\ElasticaBundle\Index\IndexManager;
 use InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Container;
+use function demosplan\DemosPlanAssessmentTableBundle\Logic\count;
 
 class AssessmentTableServiceStorage
 {
