+++ demosplan/DemosPlanCoreBundle/Logic/AssessmentTable/AssessmentTableViewMode.php
@@ -8,7 +8,7 @@
  * All rights reserved
  */
 
-namespace demosplan\DemosPlanAssessmentTableBundle\Logic;
+namespace demosplan\DemosPlanCoreBundle\Logic\AssessmentTable;
 
 use JsonSerializable;
