+++ tests/backend/core/AssessmentTable/AssessmentHandlerTest.php
@@ -11,10 +11,10 @@
 namespace Tests\Core\AssessmentTable;
 
 use DateTime;
-use demosplan\DemosPlanAssessmentTableBundle\Logic\AssessmentTableServiceOutput;
-use demosplan\DemosPlanAssessmentTableBundle\Logic\AssessmentTableViewMode;
 use demosplan\DemosPlanCoreBundle\Entity\File;
 use demosplan\DemosPlanCoreBundle\Entity\Statement\Statement;
+use demosplan\DemosPlanCoreBundle\Logic\AssessmentTable\AssessmentTableServiceOutput;
+use demosplan\DemosPlanCoreBundle\Logic\AssessmentTable\AssessmentTableViewMode;
 use demosplan\DemosPlanStatementBundle\Logic\AssessmentHandler;
 use Exception;
 use Tests\Base\FunctionalTestCase;
